+++ include/loading.h
@@ -59,6 +59,6 @@ typedef struct {
 
 Result load_entries(const char * loading_path, Entry_List_s * list);
 bool load_preview(Entry_List_s list, int * preview_offset);
-u64 load_data(char * filename, Entry_s entry, char ** buf);
+u32 load_data(char * filename, Entry_s entry, char ** buf);
 
 #endif
